# refactor(cli): shuffle imports to reduce startup time