--- a/cli/bin/capacitor
+++ b/cli/bin/capacitor
@@ -1,12 +1,12 @@
 #!/usr/bin/env node
 'use strict';
 
-var semver = require('semver');
+var satisfies = require('semver/functions/satisfies');
 var packageJson = require('../package.json');
 var currentNodeVersion = process.version.replace('v', '');
 var requiresNodeVersion = packageJson.engines.node;
 
-if (!semver.satisfies(currentNodeVersion, requiresNodeVersion)) {
+if (!satisfies(currentNodeVersion, requiresNodeVersion)) {
   console.error('ERROR: Capacitor CLI requires Node ' + requiresNodeVersion);
   process.exit(1);
 }