--- a/cli/src/common.ts
+++ b/cli/src/common.ts
@@ -1,17 +1,10 @@
-import { wordWrap } from '@ionic/cli-framework-output';
-import { copy, move, writeJSON, readFile, pathExists } from '@ionic/utils-fs';
-import { Subprocess, SubprocessError, which } from '@ionic/utils-subprocess';
-import { spawn } from 'child_process';
+import { readJSON, pathExists } from '@ionic/utils-fs';
 import { dirname, join } from 'path';
-import type { Answers, PromptObject } from 'prompts';
-import prompts from 'prompts';
-import semver from 'semver';
-import { setTimeout } from 'timers';
-import xml2js from 'xml2js';
 
 import c from './colors';
-import type { Config, ExternalConfig, PackageJson } from './definitions';
-import { output, logger } from './log';
+import type { Config, PackageJson } from './definitions';
+import { output, logger, logFatal } from './log';
+import { resolveNode } from './util/node';
 
 export type CheckFunction = () => Promise<string | null>;
 
@@ -148,107 +141,6 @@ export async function checkAppName(
   return null;
 }
 
-export async function readJSON(path: string): Promise<any> {
-  const data = await readFile(path, { encoding: 'utf-8' });
-  return JSON.parse(data);
-}
-
-export async function readXML(path: string): Promise<any> {
-  try {
-    const xmlStr = await readFile(path, { encoding: 'utf-8' });
-    try {
-      return await xml2js.parseStringPromise(xmlStr);
-    } catch (e) {
-      throw `Error parsing: ${path}, ${e.stack ?? e}`;
-    }
-  } catch (e) {
-    throw `Unable to read: ${path}`;
-  }
-}
-
-export function parseXML(xmlStr: string): any {
-  let xmlObj;
-  xml2js.parseString(xmlStr, (err: any, result: any) => {
-    if (!err) {
-      xmlObj = result;
-    }
-  });
-  return xmlObj;
-}
-
-export async function writeXML(object: any): Promise<any> {
-  return new Promise(resolve => {
-    const builder = new xml2js.Builder({
-      headless: true,
-      explicitRoot: false,
-      rootName: 'deleteme',
-    });
-    let xml = builder.buildObject(object);
-    xml = xml.replace('<deleteme>', '').replace('</deleteme>', '');
-    resolve(xml);
-  });
-}
-
-export function buildXmlElement(configElement: any, rootName: string): string {
-  const builder = new xml2js.Builder({
-    headless: true,
-    explicitRoot: false,
-    rootName: rootName,
-  });
-  return builder.buildObject(configElement);
-}
-
-export async function mergeConfig(
-  config: Config,
-  extConfig: ExternalConfig,
-): Promise<void> {
-  const oldConfig = { ...config.app.extConfig };
-
-  await writeJSON(
-    config.app.extConfigFilePath,
-    {
-      ...oldConfig,
-      ...extConfig,
-      ...{
-        plugins: extConfig.plugins ??
-          oldConfig.plugins ?? { SplashScreen: { launchShowDuration: 0 } },
-      },
-    },
-    { spaces: 2 },
-  );
-}
-
-export async function logPrompt<T extends string>(
-  msg: string,
-  prompt: PromptObject<T>,
-): Promise<Answers<T>> {
-  logger.log({
-    msg: `${c.input('[?]')} ${wordWrap(msg, { indentation: 4 })}`,
-    logger,
-    format: false,
-  });
-  return prompts(prompt, { onCancel: () => process.exit(1) });
-}
-
-export function logSuccess(msg: string): void {
-  logger.msg(`${c.success('[success]')} ${msg}`);
-}
-
-export function logFatal(msg: string): never {
-  logger.error(msg);
-  return process.exit(1);
-}
-
-export async function isInstalled(command: string): Promise<boolean> {
-  try {
-    await which(command);
-  } catch (e) {
-    return false;
-  }
-
-  return true;
-}
-
 export async function wait(time: number): Promise<void> {
   return new Promise(resolve => setTimeout(resolve, time));
 }
@@ -257,6 +149,7 @@ export async function runPlatformHook(
   platformDir: string,
   hook: string,
 ): Promise<void> {
+  const { spawn } = await import('child_process');
   const pkg = await readJSON(join(platformDir, 'package.json'));
   const cmd = pkg.scripts[hook];
 
@@ -283,57 +176,6 @@ export async function runPlatformHook(
   });
 }
 
-export interface RunCommandOptions {
-  cwd?: string;
-}
-
-export async function runCommand(
-  command: string,
-  args: readonly string[],
-  options: RunCommandOptions = {},
-): Promise<string> {
-  const p = new Subprocess(command, args, options);
-
-  try {
-    return await p.output();
-  } catch (e) {
-    if (e instanceof SubprocessError) {
-      // old behavior of just throwing the stdout/stderr strings
-      throw e.output ? e.output : e.code;
-    }
-
-    throw e;
-  }
-}
-
-export async function runNativeRun(
-  args: readonly string[],
-  options: RunCommandOptions = {},
-): Promise<string> {
-  const p = resolveNode(
-    __dirname,
-    dirname('native-run/package'),
-    'bin/native-run',
-  );
-
-  if (!p) {
-    logFatal(`${c.input('native-run')} not found.`);
-  }
-
-  return await runCommand(p, args, options);
-}
-
-export async function getCommandOutput(
-  command: string,
-  args: readonly string[],
-): Promise<string | null> {
-  try {
-    return (await runCommand(command, args)).trim();
-  } catch (e) {
-    return null;
-  }
-}
-
 export interface RunTaskOptions {
   spinner?: boolean;
 }
@@ -352,20 +194,6 @@ export async function runTask<T>(
   } catch (e) {
     chain.fail();
     throw e;
-  }
-}
-
-export async function copyTemplate(src: string, dst: string): Promise<void> {
-  await copy(src, dst);
-  await renameGitignore(dst);
-}
-
-export async function renameGitignore(dst: string): Promise<void> {
-  // npm renames .gitignore to something else, so our templates
-  // have .gitignore as gitignore, we need to rename it here.
-  const gitignorePath = join(dst, 'gitignore');
-  if (await pathExists(gitignorePath)) {
-    await move(gitignorePath, join(dst, '.gitignore'));
   }
 }
 
@@ -499,8 +327,10 @@ export async function promptForPlatform(
   promptMessage: string,
   selectedPlatformName?: string,
 ): Promise<string> {
+  const { prompt } = await import('prompts');
+
   if (!selectedPlatformName) {
-    const answers = await prompts(
+    const answers = await prompt(
       [
         {
           type: 'select',
@@ -529,15 +359,26 @@ export async function promptForPlatform(
   return platformName;
 }
 
+export interface PlatformTarget {
+  id: string;
+  platform: string;
+  virtual: boolean;
+  sdkVersion: string;
+  name?: string;
+  model?: string;
+}
+
 export async function promptForPlatformTarget(
   targets: PlatformTarget[],
   selectedTarget?: string,
 ): Promise<PlatformTarget> {
+  const { prompt } = await import('prompts');
+
   if (!selectedTarget) {
     if (targets.length === 1) {
       return targets[0];
     } else {
-      const answers = await prompts(
+      const answers = await prompt(
         [
           {
             type: 'select',
@@ -569,6 +410,14 @@ export async function promptForPlatformTarget(
   return target;
 }
 
+export function getPlatformTargetName(target: PlatformTarget): string {
+  return `${target.name ?? target.model ?? target.id ?? '?'}${
+    target.virtual
+      ? ` (${target.platform === 'ios' ? 'simulator' : 'emulator'})`
+      : ''
+  }`;
+}
+
 export async function getAddedPlatforms(config: Config): Promise<string[]> {
   const platforms: string[] = [];
 
@@ -589,8 +438,10 @@ export async function checkPlatformVersions(
   config: Config,
   platform: string,
 ): Promise<void> {
+  const semver = await import('semver');
   const coreVersion = await getCoreVersion(config);
   const platformVersion = await getCapacitorPackageVersion(config, platform);
+
   if (
     semver.diff(coreVersion, platformVersion) === 'minor' ||
     semver.diff(coreVersion, platformVersion) === 'major'
@@ -606,35 +457,6 @@ export async function checkPlatformVersions(
         )}`,
     );
   }
-}
-
-export interface PlatformTarget {
-  id: string;
-  platform: string;
-  virtual: boolean;
-  sdkVersion: string;
-  name?: string;
-  model?: string;
-}
-
-export async function getPlatformTargets(
-  platformName: string,
-): Promise<PlatformTarget[]> {
-  const output = await runNativeRun([platformName, '--list', '--json']);
-  const parsedOutput = JSON.parse(output);
-
-  return [
-    ...parsedOutput.devices.map((t: any) => ({ ...t, virtual: false })),
-    ...parsedOutput.virtualDevices.map((t: any) => ({ ...t, virtual: true })),
-  ];
-}
-
-export function getPlatformTargetName(target: PlatformTarget): string {
-  return `${target.name ?? target.model ?? target.id ?? '?'}${
-    target.virtual
-      ? ` (${target.platform === 'ios' ? 'simulator' : 'emulator'})`
-      : ''
-  }`;
 }
 
 export function resolvePlatform(
@@ -671,15 +493,4 @@ export function resolvePlatform(
   }
 
   return null;
-}
-
-export function resolveNode(
-  root: string,
-  ...pathSegments: string[]
-): string | null {
-  try {
-    return require.resolve(pathSegments.join('/'), { paths: [root] });
-  } catch (e) {
-    return null;
-  }
 }