--- a/cli/src/config.ts
+++ b/cli/src/config.ts
@@ -3,7 +3,6 @@ import Debug from 'debug';
 import { dirname, join, resolve } from 'path';
 
 import c from './colors';
-import { logFatal, resolveNode, runCommand } from './common';
 import type {
   AndroidConfig,
   AppConfig,
@@ -14,8 +13,9 @@ import type {
   WebConfig,
 } from './definitions';
 import { OS } from './definitions';
+import { logFatal } from './log';
 import { tryFn } from './util/fn';
-import { requireTS } from './util/node';
+import { resolveNode, requireTS } from './util/node';
 
 const debug = Debug('capacitor:config');
 
@@ -250,6 +250,8 @@ async function determineAndroidStudioPath(os: OS): Promise<string> {
     case OS.Mac:
       return '/Applications/Android Studio.app';
     case OS.Windows: {
+      const { runCommand } = await import('./util/subprocess');
+
       let p = 'C:\\Program Files\\Android\\Android Studio\\bin\\studio64.exe';
 
       try {