--- a/cli/src/ios/open.ts
+++ b/cli/src/ios/open.ts
@@ -1,8 +1,9 @@
 import open from 'open';
 
 import c from '../colors';
-import { wait, logFatal } from '../common';
+import { wait } from '../common';
 import type { Config } from '../definitions';
+import { logFatal } from '../log';
 
 import { findXcodePath } from './common';