--- a/cli/src/log.ts
+++ b/cli/src/log.ts
@@ -5,11 +5,12 @@ import {
   TTYOutputStrategy,
   createDefaultLogger,
 } from '@ionic/cli-framework-output';
+import type { Answers, PromptObject } from 'prompts';
 
-import colors from './colors';
+import c from './colors';
 import { isInteractive } from './util/term';
 
-const options = { colors, stream: process.stdout };
+const options = { colors: c, stream: process.stdout };
 
 export const output = isInteractive()
   ? new TTYOutputStrategy(options)
@@ -20,10 +21,35 @@ export const logger = createDefaultLogger({
   formatterOptions: {
     titleize: false,
     tags: new Map<LoggerLevelWeight, string>([
-      [LOGGER_LEVELS.DEBUG, colors.log.DEBUG('[debug]')],
-      [LOGGER_LEVELS.INFO, colors.log.INFO('[info]')],
-      [LOGGER_LEVELS.WARN, colors.log.WARN('[warn]')],
-      [LOGGER_LEVELS.ERROR, colors.log.ERROR('[error]')],
+      [LOGGER_LEVELS.DEBUG, c.log.DEBUG('[debug]')],
+      [LOGGER_LEVELS.INFO, c.log.INFO('[info]')],
+      [LOGGER_LEVELS.WARN, c.log.WARN('[warn]')],
+      [LOGGER_LEVELS.ERROR, c.log.ERROR('[error]')],
     ]),
   },
 });
+
+export async function logPrompt<T extends string>(
+  msg: string,
+  promptObject: PromptObject<T>,
+): Promise<Answers<T>> {
+  const { wordWrap } = await import('@ionic/cli-framework-output');
+  const { prompt } = await import('prompts');
+
+  logger.log({
+    msg: `${c.input('[?]')} ${wordWrap(msg, { indentation: 4 })}`,
+    logger,
+    format: false,
+  });
+
+  return prompt(promptObject, { onCancel: () => process.exit(1) });
+}
+
+export function logSuccess(msg: string): void {
+  logger.msg(`${c.success('[success]')} ${msg}`);
+}
+
+export function logFatal(msg: string): never {
+  logger.error(msg);
+  return process.exit(1);
+}