… cli/src/tasks/open.ts
… cli/src/tasks/run.ts
… cli/src/tasks/serve.ts
… cli/src/tasks/sync.ts
… cli/src/tasks/update.ts
… cli/src/util/native-run.ts
… cli/src/util/node.ts
… cli/src/util/subprocess.ts
… cli/src/util/template.ts
… cli/src/util/xml.ts
… cli/src/web/copy.ts
… cli/test/util.ts