Using the following code after config is loaded in the `run()` function of `cli/src/index.ts`:

```js
console.log(Object.keys(require.cache).length);
```

as well as `time npx cap --version` to test startup time.

**Benchmarks**

These are the numbers I'm getting on my machine (MacBook Pro (16-inch, 2019) / 2.4 GHz 8-Core Intel Core i9 / 32 GB 2667 MHz DDR4):

`main`
- 280 modules in cache
- 0.42s startup
- package: 141 files (500.7 kB, 829.7 kB unpacked)

`cli-bundle` (https://github.com/ionic-team/capacitor/pull/3797)
- 241 modules in cache
- 0.40s startup
- package: 102 files (497.3 kB, 807.7 kB unpacked)

`import-refactor` (this PR)
- 100 modules in cache
- 0.32s startup 
- package: 145 files (501.1 kB, 832.9 kB unpacked)

**Results**

This has ~100ms faster startup than `cli-bundle` while maintaining the simple build process and useful dist stack traces and while adding little to no maintenance overhead. To reduce the number of files published we can implement https://github.com/ionic-team/capacitor/issues/3290 (wouldn't reduce the packaged size much, but would offer additional benefits).

alternative to https://github.com/ionic-team/capacitor/pull/3797